feat(secrets): add option to get and set the secret store (#4707)
